exercises/04-automate-security-checks-python/requirements.txt
boto3
google-cloud-storage
azure-identity
azure-mgmt-storage